Making what "RIBs" is more clear on the README
Reading some early feedback, it seems that it's not clear to some people meeting the architecture the first time what "RIBs" stands for. Sending a suggestion to make this clear, as well as linking the original announcement post (the latter we can update if we will have a more up-to-date one).